+++ x/superfluid/keeper/epoch.go
@@ -3,11 +3,9 @@ package keeper
 import (
 	"errors"
 	"fmt"
-	"time"
-
-	sdkerrors "cosmossdk.io/errors"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
+
 	distributiontypes "github.com/cosmos/cosmos-sdk/x/distribution/types"
 
 	"github.com/osmosis-labs/osmosis/osmoutils"
@@ -151,23 +149,10 @@ func (k Keeper) UpdateOsmoEquivalentMultipliers(ctx sdk.Context, asset types.Sup
 			return k.updateConcentratedOsmoEquivalentMultiplier(cacheCtx, asset, newEpochNumber)
 		})
 	} else if asset.AssetType == types.SuperfluidAssetTypeNative {
-		bondDenom, err := k.sk.BondDenom(ctx)
-		if err != nil {
-			return err
-		}
-		if asset.Denom == bondDenom {
-			// The bond denom should be locked via x/lockup and not superfluid
-			return errors.New("osmo should not be a superfluid asset. It can be staked natively")
-		}
-		// Twap price is calculated from the last 5 minutes
-		startTime := ctx.BlockTime().Add(-5 * time.Minute) // TODO: make this configurable
-		price, err := k.twapk.UnsafeGetMultiPoolArithmeticTwapToNow(ctx, asset.PriceRoute, asset.Denom, bondDenom, startTime)
-		if err != nil {
-			return sdkerrors.Wrap(err, "failed to get twap price")
-		}
-		k.SetOsmoEquivalentMultiplier(ctx, newEpochNumber, asset.Denom, price)
+		// TODO: Consider deleting superfluid asset type native
+		k.Logger(ctx).Error("unsupported superfluid asset type")
+		return errors.New("SuperfluidAssetTypeNative is unsupported")
 	}
-
 	return nil
 }