+++ x/superfluid/keeper/stake.go
@@ -152,16 +152,6 @@ func (k Keeper) validateLockForSF(lock *lockuptypes.PeriodLock, sender string) e
 	return nil
 }
 
-func (k Keeper) ValidateNativeAsset(asset types.SuperfluidAsset) error {
-	if asset.AssetType == types.SuperfluidAssetTypeNative {
-		if len(asset.PriceRoute) == 0 ||
-			strings.TrimSpace(asset.PriceRoute[0].TokenOutDenom) == "" {
-			return errorsmod.Wrap(types.ErrNonSuperfluidAsset, "asset is not properly configured for superfluid staking: missing or empty price route")
-		}
-	}
-	return nil
-}
-
 // validateLockForSFDelegate runs the following sanity checks on the lock:
 // - the sender is the owner of the lock
 // - the lock is consisted of a single coin
@@ -178,13 +168,8 @@ func (k Keeper) validateLockForSFDelegate(ctx sdk.Context, lock *lockuptypes.Per
 	denom := lock.Coins[0].Denom
 
 	// ensure that the locks underlying denom is for an existing superfluid asset
-	asset, err := k.GetSuperfluidAsset(ctx, denom)
+	_, err = k.GetSuperfluidAsset(ctx, denom)
 	if err != nil {
-		return err
-	}
-
-	// ensure that the asset is properly configured
-	if err := k.ValidateNativeAsset(asset); err != nil {
 		return err
 	}
 
@@ -612,7 +597,7 @@ func (k Keeper) IterateDelegations(context context.Context, delegator sdk.AccAdd
 		}
 
 		// get osmo-equivalent token amount
-		amount, err := k.GetSuperfluidOSMOTokensIfNonNative(ctx, interim.Denom, coin.Amount)
+		amount, err := k.GetSuperfluidOSMOTokens(ctx, interim.Denom, coin.Amount)
 		if err != nil {
 			ctx.Logger().Error("failed to get osmo equivalent of token", "Denom", interim.Denom, "Amount", coin.Amount, "Error", err)
 			return err