+++ x/superfluid/types/expected_keepers.go
@@ -4,8 +4,6 @@ import (
 	context "context"
 	"time"
 
-	poolmanagertypes "github.com/osmosis-labs/osmosis/v25/x/poolmanager/types"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	stakingtypes "github.com/cosmos/cosmos-sdk/x/staking/types"
 
@@ -142,8 +140,4 @@ type PoolManagerKeeper interface {
 
 type ValSetPreferenceKeeper interface {
 	DelegateToValidatorSet(ctx sdk.Context, delegatorAddr string, coin sdk.Coin) error
-}
-
-type TwapKeeper interface {
-	UnsafeGetMultiPoolArithmeticTwapToNow(ctx sdk.Context, route []*poolmanagertypes.SwapAmountInRoute, baseAssetDenom string, quoteAssetDenom string, startTime time.Time) (osmomath.Dec, error)
 }